refactor: extract daily planner and timeline into modular components
