+++ package.json
@@ -4,18 +4,18 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "dev:open": "./scripts/launch.sh dev",
-    "balance:start": "./scripts/balance-server.sh start",
-    "balance:stop": "./scripts/balance-server.sh stop",
-    "balance:restart": "./scripts/balance-server.sh restart",
-    "balance:open": "./scripts/balance-server.sh open",
-    "balance:status": "./scripts/balance-server.sh status",
-    "balance:logs": "./scripts/balance-server.sh logs",
-    "balance:install": "./scripts/balance-server.sh install",
-    "balance:uninstall": "./scripts/balance-server.sh uninstall",
+    "dev:open": "./scripts/ctl.sh open dev --fg",
+    "balance:start": "./scripts/ctl.sh start",
+    "balance:stop": "./scripts/ctl.sh stop",
+    "balance:restart": "./scripts/ctl.sh restart",
+    "balance:open": "./scripts/ctl.sh open",
+    "balance:status": "./scripts/ctl.sh status",
+    "balance:logs": "./scripts/ctl.sh logs",
+    "balance:install": "./scripts/ctl.sh install",
+    "balance:uninstall": "./scripts/ctl.sh uninstall",
     "build": "next build",
     "start": "next start",
-    "start:open": "./scripts/launch.sh start",
+    "start:open": "./scripts/ctl.sh open prod --fg",
     "lint": "eslint"
   },
   "dependencies": {
