+++ src/components/timeline/timelineUtils.ts
@@ -1,0 +1,13 @@
+export function formatTimelineDate(date: string): string {
+  return new Intl.DateTimeFormat("en-US", {
+    weekday: "short",
+    month: "short",
+    day: "numeric",
+  }).format(new Date(`${date}T12:00:00`));
+}
+
+export function getTimelineKicker(date: string, today: string, tomorrow: string): string {
+  if (date === today) return "Today";
+  if (date === tomorrow) return "Tomorrow";
+  return date < today ? "Past" : "Upcoming";
+}
