+++ src/hooks/useAuth.tsx
@@ -28,18 +28,27 @@ export function AuthProvider({ children }: { children: ReactNode }) {
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
-    supabase.auth.getSession().then(({ data: { session } }) => {
-      setSession(session);
-      setUser(session?.user ?? null);
-      setLoading(false);
-    });
+    let initialDone = false;
 
     const {
       data: { subscription },
     } = supabase.auth.onAuthStateChange((_event, session) => {
       setSession(session);
       setUser(session?.user ?? null);
-      setLoading(false);
+      if (!initialDone) {
+        initialDone = true;
+        setLoading(false);
+      }
+    });
+
+    // Fallback: if onAuthStateChange doesn't fire quickly, resolve via getSession
+    supabase.auth.getSession().then(({ data: { session } }) => {
+      if (!initialDone) {
+        initialDone = true;
+        setSession(session);
+        setUser(session?.user ?? null);
+        setLoading(false);
+      }
     });
 
     return () => subscription.unsubscribe();
